style: Refactor carousel component styles and structure
- Removed outdated styles from page.module.css to streamline the CSS structure.
- Introduced a new EmblaCarousel.module.css file for the carousel component, enhancing layout and responsiveness.
- Updated import paths in related components to reflect the new CSS file location.
- Improved accessibility features and responsive design for better user experience across devices.
