+++ src/components/carousel/EmblaCarousel.module.css
@@ -1,4 +1,4 @@
-@import '../../../styles/breakpoints.css';
+@import '../../styles/breakpoints.css';
 
 /* Carousel Container */
 .embla {
@@ -83,7 +83,7 @@
 
 /* Button Base Styles */
 .button {
-  -webkit-tap-highlight-color: rgba(var(--text-high-contrast-rgb-value), 0.5);
+  -webkit-tap-highlight-color: rgba(255, 255, 255, 0.5);
   -webkit-appearance: none;
   appearance: none;
   background-color: transparent;
@@ -148,7 +148,6 @@
 .dot {
   width: 2rem;
   height: 2rem;
-  border-radius: 50%;
   border-radius: 50%;
   background: rgba(255, 255, 255, 0.1);
   -webkit-backdrop-filter: blur(3px);
